examine service: split into controller and service
